Completing the mapping function
Successfully computing the mapping function to map the location of the white colored paper being tracked by Processing to the location over the sandbox map in Unity.
If anything changes, the mapping formulas will need to be changed as well. This includes resizing or moving windows around.

Problem now: Mouse click doesn't click through programs. Unity will refresh in the background, but it doesn't detect mouse clicks that are made in different windows. Processing will not run in the background. So have to come up with a solution to make the mouse click through a program, so it can click and drag through Processing into Unity so it can move the rain.